pkg/rhcos: Add a stub package for calculating RHCOS AMIs
We currently calculate CoreOS Container Linux AMIs in Terraform
(installer/modules/aws/ami), but with the coming shift to RHCOS and Go
asset generation, it would be nice to calculate them in Go.  This
commit adds a package to do so.  Currently it's enough of a stub to
support testing, with the value based on [1]:

  {
    "HVM":"ami-06d864b4154214132",
    "SnapshotID":"snap-07b63a4c2f8869c15",
    "S3Object":"s3://openshift-qe-images/rhcos/cloud/rhcos-4.0.5122-aws.vmdk"
  }

from [2].  Once we get a public version of [2] (plans in [3]), we can
replace the stub in this package with something that works for more
regions and channels.

[1]: https://github.com/openshift/release/pull/1344#issue-212549406
[2]: http://aos-ostree.rhev-ci-vms.eng.rdu2.redhat.com/rhcos/images/aws-us-east-1-tested.json
[3]: https://github.com/openshift/release/pull/1344#issuecomment-418449544